refactor(utils): remove unused helpers and inline single-use highlight
Closes #333.

- worker/utils/types.ts: drop unused isStringArray / NonEmptyArray
  (makeOneOf retained — used by 14-yaml-feed-config validator)
- client/utils/highlight.tsx: deleted (used by only ArticleCard)
- ArticleCard.tsx: inline highlight as local helper (single caller, no export)
- tests/client/highlight.test.tsx: removed in favor of behavior tests on ArticleCard
- tests/client/ArticleCard.test.tsx: 3 new tests verifying highlight behavior

Aligns with new utils placement guideline in 01-typescript.md.
