+++ apps/web/client/components/ArticleCard.tsx
@@ -1,10 +1,25 @@
+import type { ReactNode } from "react";
 import { useEffect, useRef, useState } from "react";
 import { useReadState } from "../hooks/useReadState";
 import { useStarredState } from "../hooks/useStarredState";
 import type { Article, FeedCategory } from "../types/api";
-import { highlight } from "../utils/highlight";
 import { BookmarkButton } from "./BookmarkButton";
 import { ShareButtons } from "./ShareButtons";
+
+function escapeRegex(s: string): string {
+  return s.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
+
+function highlight(text: string, q: string): ReactNode {
+  if (!q) return text;
+  const re = new RegExp(`(${escapeRegex(q)})`, "ig");
+  const parts = text.split(re);
+  // split with a capture group alternates: non-match, match, non-match, ...
+  return parts.map((part, i) =>
+    // even indices are non-matches, odd indices are the captured matches
+    i % 2 === 1 ? <mark key={i}>{part}</mark> : part,
+  );
+}
 
 interface Props {
   article: Article;
